+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a new method `setProvider(Provider provider)` in `PcscPluginFactoryBuilder` to allow the use of a custom PC/SC
+  provider.
 
 ## [2.3.1] - 2024-10-01
 ### Fixed